+++ src/sys/os/linux/fns/syscalls/arm.rs
@@ -14,8 +14,9 @@ use core::{
 pub unsafe fn linux_sys_exit(status: c_int) -> ! {
     unsafe {
         asm!(
+            "mov rax, {EXIT}",
             "swi 0",
-            in("r8") SYS::EXIT,
+            EXIT = const SYS::EXIT,
             in("r0") status,
             options(noreturn)
         );
@@ -25,100 +26,112 @@ pub unsafe fn linux_sys_exit(status: c_int) -> ! {
 #[doc = include_str!("./doc/Sys_read.md")]
 #[must_use]
 pub unsafe fn linux_sys_read(fd: c_int, buf: *mut u8, count: usize) -> isize {
-    let r0;
+    let result;
     unsafe {
         asm!(
+            "mov rax, {READ}",
             "swi 0",
-            inlateout("r8") SYS::READ => r0,
+            READ = const SYS::READ,
             in("r0") fd,
             in("r1") buf,
             in("r2") count,
-            options(nostack, preserves_flags)
+            lateout("r7") result,
+            options(nostack)
         );
     }
-    r0
+    result
 }
 
 #[doc = include_str!("./doc/Sys_write.md")]
 #[must_use]
 pub unsafe fn linux_sys_write(fd: c_int, buf: *const u8, count: usize) -> isize {
-    let r0;
+    let result;
     unsafe {
         asm!(
+            "mov rax, {WRITE}",
             "swi 0",
-            inlateout("r8") SYS::WRITE => r0,
+            WRITE = const SYS::WRITE,
             in("r0") fd,
             in("r1") buf,
             in("r2") count,
-            options(nostack, preserves_flags)
+            lateout("r7") result,
+            options(nostack)
         );
     }
-    r0
+    result
 }
 
 #[doc = include_str!("./doc/Sys_nanosleep.md")]
 #[must_use]
 pub unsafe fn linux_sys_nanosleep(req: *const LinuxTimespec, rem: *mut LinuxTimespec) -> isize {
-    let r0;
+    let result;
     unsafe {
         asm!(
+            "mov rax, {NANOSLEEP}",
             "swi 0",
-            inlateout("r8") SYS::NANOSLEEP => r0,
+            NANOSLEEP = const SYS::NANOSLEEP,
             in("r0") req,
             in("r1") rem,
             lateout("r2") _,
-            options(nostack, preserves_flags)
+            lateout("r7") result,
+            options(nostack)
         );
     }
-    r0
+    result
 }
 
 #[doc = include_str!("./doc/Sys_ioctl.md")]
 #[must_use]
 pub unsafe fn linux_sys_ioctl(fd: c_int, request: c_ulong, argp: *mut u8) -> isize {
-    let r0;
+    let result;
     unsafe {
         asm!(
+            "mov rax, {IOCTL}",
             "swi 0",
-            inlateout("r8") SYS::IOCTL => r0,
+            IOCTL = const SYS::IOCTL,
             in("r0") fd,
             in("r1") request,
             in("r2") argp,
-            options(nostack, preserves_flags)
+            lateout("r7") result,
+            options(nostack)
         );
     }
-    r0
+    result
 }
 
 #[doc = include_str!("./doc/Sys_getrandom.md")]
 #[must_use]
 pub unsafe fn linux_sys_getrandom(buffer: *mut u8, size: usize, flags: c_uint) -> isize {
-    let r0;
+    let result;
     unsafe {
         asm!(
+            "mov rax, {GETRANDOM}",
             "swi 0",
-            inlateout("r8") SYS::GETRANDOM => r0,
+            GETRANDOM = const SYS::GETRANDOM,
             in("r0") buffer,
             in("r1") size,
             in("r2") flags,
-            options(nostack, preserves_flags)
+            lateout("r7") result,
+            options(nostack)
         );
     }
-    r0
+    result
 }
 
 #[doc = include_str!("./doc/Sys_getpid.md")]
 #[must_use]
 pub unsafe fn linux_sys_getpid() -> i32 {
-    let r0: isize;
+    let result: isize;
     unsafe {
         asm!(
+            "mov rax, {GETPID}",
             "swi 0",
-            inlateout("r8") SYS::GETPID => r0,
-            options(nostack, preserves_flags)
+            GETPID = const SYS::GETPID,
+            lateout("r7") result,
+            options(nostack)
         );
     }
-    r0 as i32
+    result as i32
 }
 
 #[doc = include_str!("./doc/Sys_rt_sigaction.md")]
@@ -129,17 +142,19 @@ pub unsafe fn linux_sys_rt_sigaction(
     oact: *mut LinuxSigaction,
     sigsetsize: usize,
 ) -> isize {
-    let r0;
+    let result;
     unsafe {
         asm!(
+            "mov rax, {RT_SIGACTION}",
             "swi 0",
-            inlateout("r8") SYS::RT_SIGACTION => r0,
+            RT_SIGACTION = const SYS::RT_SIGACTION,
             in("r0") sig,
             in("r1") act,
             in("r2") oact,
             in("r3") sigsetsize,
-            options(nostack, preserves_flags)
+            lateout("r7") result,
+            options(nostack)
         );
     }
-    r0
+    result
 }