+++ src/sys/os/linux/fns/syscalls/riscv.rs
@@ -13,8 +13,9 @@ use core::{
 pub unsafe fn linux_sys_exit(status: c_int) -> ! {
     unsafe {
         asm!(
+            "li a7, {EXIT}",
             "ecall",
-            in("a7") SYS::EXIT,
+            EXIT = const SYS::EXIT,
             in("a0") status,
             options(noreturn)
         );
@@ -24,100 +25,112 @@ pub unsafe fn linux_sys_exit(status: c_int) -> ! {
 #[doc = include_str!("./doc/Sys_read.md")]
 #[must_use]
 pub unsafe fn linux_sys_read(fd: c_int, buf: *mut u8, count: usize) -> isize {
-    let r0;
+    let result;
     unsafe {
         asm!(
+            "li a7, {READ}",
             "ecall",
-            inlateout("a7") SYS::READ => r0,
+            READ = const SYS::READ,
             in("a0") fd,
             in("a1") buf,
             in("a2") count,
-            options(nostack, preserves_flags)
+            lateout("a7") result,
+            options(nostack)
         );
     }
-    r0
+    result
 }
 
 #[doc = include_str!("./doc/Sys_write.md")]
 #[must_use]
 pub unsafe fn linux_sys_write(fd: c_int, buf: *const u8, count: usize) -> isize {
-    let r0;
+    let result;
     unsafe {
         asm!(
+            "li a7, {WRITE}",
             "ecall",
-            inlateout("a7") SYS::WRITE => r0,
+            WRITE = const SYS::WRITE,
             in("a0") fd,
             in("a1") buf,
             in("a2") count,
-            options(nostack, preserves_flags)
-        );
+            lateout("a7") result,
+            options(nostack)
+        )
     }
-    r0
+    result
 }
 
 #[doc = include_str!("./doc/Sys_nanosleep.md")]
 #[must_use]
 pub unsafe fn linux_sys_nanosleep(req: *const LinuxTimespec, rem: *mut LinuxTimespec) -> isize {
-    let r0;
+    let result;
     unsafe {
         asm!(
+            "li a7, {NANOSLEEP}",
             "ecall",
-            inlateout("a7") SYS::NANOSLEEP => r0,
-            in("a0") fd,
-            in("a1") buf,
+            NANOSLEEP = const SYS::NANOSLEEP,
+            in("a0") req,
+            in("a1") rem,
+            lateout("a7") result,
             lateout("a2") _,
-            options(nostack, preserves_flags)
-        );
+            options(nostack)
+        )
     }
-    r0
+    result
 }
 
 #[doc = include_str!("./doc/Sys_ioctl.md")]
 #[must_use]
 pub unsafe fn linux_sys_ioctl(fd: c_int, request: c_ulong, argp: *mut u8) -> isize {
-    let r0;
+    let result;
     unsafe {
         asm!(
+            "li a7, {IOCTL}",
             "ecall",
-            inlateout("a7") SYS::IOCTL => r0,
+            IOCTL = const SYS::IOCTL,
             in("a0") fd,
             in("a1") request,
             in("a2") argp,
-            options(nostack, preserves_flags)
+            lateout("a7") result,
+            options(nostack)
         );
     }
-    r0
+    result
 }
 
 #[doc = include_str!("./doc/Sys_getrandom.md")]
 #[must_use]
 pub unsafe fn linux_sys_getrandom(buffer: *mut u8, size: usize, flags: c_uint) -> isize {
-    let r0;
+    let result;
     unsafe {
         asm!(
+            "li a7, {GETRANDOM}",
             "ecall",
-            inlateout("a7") SYS::GETRANDOM => r0,
+            GETRANDOM = const SYS::GETRANDOM,
             in("a0") buffer,
             in("a1") size,
             in("a2") flags,
-            options(nostack, preserves_flags)
+            lateout("a7") result,
+            options(nostack)
         );
     }
-    r0
+    result
 }
 
 #[doc = include_str!("./doc/Sys_getpid.md")]
 #[must_use]
 pub unsafe fn linux_sys_getpid() -> i32 {
-    let r0: isize;
+    let result: isize;
     unsafe {
         asm!(
+            "li a7, {GETPID}",
             "ecall",
-            inlateout("a7") SYS::GETPID => r0,
-            options(nostack, preserves_flags)
+            GETPID = const SYS::GETPID,
+            lateout("a7") result,
+            options(nostack)
         );
     }
-    r0 as i32
+    result as i32
 }
 
 #[doc = include_str!("./doc/Sys_rt_sigaction.md")]
@@ -128,17 +141,19 @@ pub unsafe fn linux_sys_rt_sigaction(
     oact: *mut LinuxSigaction,
     sigsetsize: usize,
 ) -> isize {
-    let r0;
+    let result;
     unsafe {
         asm!(
+            "li a7, {RT_SIGACTION}",
             "ecall",
-            inlateout("a7") SYS::RT_SIGACTION => r0,
+            RT_SIGACTION = const SYS::RT_SIGACTION,
             in("a0") sig,
             in("a1") act,
             in("a2") oact,
             in("a3") sigsetsize,
-            options(nostack, preserves_flags)
-        );
+            lateout("a7") result,
+            options(nostack)
+        )
     }
-    r0
+    result
 }