+++ src/sys/os/linux/mod.rs
@@ -89,6 +89,8 @@ pub mod thread {
             target_arch = "x86",
             target_arch = "arm",
             target_arch = "aarch64",
+            target_arch = "riscv32",
+            target_arch = "riscv64"
         ),
         feature = "unsafe_syscall",
         not(miri),